Don't dynamically generate version in setup.py
Running setup.py grabs the version from lib/iris/__init__.py (if it
exists; otherwise the version would be `None`).

Some installation tools that parse setup.py (but do not execute it),
will not find a proper version, just a function call (which will not
be executed).

The version value should be fixed in setup.py, and then possibly added
to the package upon building or installing it, not the other way
around.

The current solution is simpler, but requires double bookkeeping in
that the version value exists twice in the code: once in the package,
and once in setup.py.

Admittedly, the corresponding installation/build tool should just call
`python setup.py --version` and use that instead.